+++ src/jobs/closeExpiredAppsCheck.ts
@@ -1,0 +1,6 @@
+// Check for applications that have reached expiry date + DAYS_POST_EXPIRY
+// if they are not already in REVIEW state, they will transition to CLOSED
+export default async function () {
+  // TODO: implement
+  return;
+}
