+++ src/jobs/firstExpiryNotification.ts
@@ -1,0 +1,5 @@
+// 1st notification for applications entering renewal period (DAYS_TO_EXPIRY_1)
+export default async function () {
+  // TODO: implement
+  return;
+}
