+++ src/jobs/secondExpiryNotification.ts
@@ -1,0 +1,5 @@
+// 2nd notification for applications that have not begun renewal process (DAYS_TO_EXPIRY_2)
+export default async function () {
+  // TODO: implement
+  return;
+}
